pyproject.toml
[project]
dependencies = [
  "torch>=2.0.0",
  "torchvision>=0.15.0",
  "torchaudio>=2.0.0",
  "ultralytics>=8.0.0",
  "ffmpeg-python>=0.2.0",
  "scipy>=1.10.0",
  "scikit-learn>=1.2.0",
  "tqdm>=4.65.0",
  "scenedetect>=0.6.0",
  "opencv-python>=4.7.0",
  "python-speech-features>=0.6",
  "gdown>=4.7.0",
  "yt-dlp>=2023.0.0",
  "pandas>=2.3.3"
]
description = "Active speaker detection in videos using YOLO face detection and TalkNet audio-visual classification"
name = "active-speaker-detection"
readme = "README.md"
requires-python = ">=3.12,<3.13"
version = "0.1.0"

[project.optional-dependencies]
dev = [
  "flake8>=7.2.0",
  "pre-commit>=4.2.0",
  "toml-sort>=0.24.2"
]

[tool.uv]
extra-index-url = ["https://download.pytorch.org/whl/cu124"]
package = false

[tool.uv.sources]
torch = {index = "pytorch-cu124"}
torchaudio = {index = "pytorch-cu124"}
torchvision = {index = "pytorch-cu124"}

[[tool.uv.index]]
explicit = true
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
